feat: updated tables and made db call with get_balance

+++ commands/src/balance_commands.rs
@@ -2,14 +2,14 @@ use serde::Deserialize;
 
 #[derive(Deserialize)]
 pub struct GetBalanceCommand {
-    pub user_id: i64, 
-    pub group_id: i64,
+    pub user_id: i32,
+    pub group_id: i32,
 }
 
 #[derive(Deserialize)]
 pub struct UpdateBalancesCommand {
-    pub submitter_id: i64, 
-    pub group_id: i64,
-    pub user_ids: Vec<i64>,
-    pub amt: f64,
+    pub submitter_id: i32, 
+    pub group_id: i32,
+    pub user_ids: Vec<i32>,
+    pub amt: f32,
 }
