+++ models/src/balance_model.rs
@@ -2,14 +2,14 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Balance {
-    pub balance_id: i64,
-    pub user_id: i64,
-    pub group_id: i64,
-    pub amt: f64,
+    pub balance_id: i32,
+    pub user_id: i32,
+    pub group_id: i32,
+    pub amt: f32,
 }
 
 impl Balance {
-    pub fn new(user_id: i64, group_id: i64, amt: f64) -> Balance {
+    pub fn new(user_id: i32, group_id: i32, amt: f32) -> Balance {
         Balance {
             balance_id: 1,
             user_id,
